requirements.txt
anyio==3.6.2
asgiref==3.6.0
channels==4.0.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
django-ratelimit==4.0.0
djangorestframework==3.14.0
h11==0.14.0
httptools==0.5.0
idna==3.4
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
ratelimit==2.2.1
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
